[export][dynamo] Fallback to slowpath for MultiHeadAttention for strict export
This is undesirable but MultiHeadAttention has some weirdness to it
where if we are make_fx tracing (_is_make_fx_tracing), it goes to slow
path. This weirdness leads to different behavior between non-strict
export and compile.

For non-strict export, _is_make_fx_tracing path is True and the captured
graph is slow path. But for compile, for Dynamo, the _is_make_fx_tracing
cond is False, leading to fast graph.

This inherent divergence leads to test failures in executorch. We should
figure out why _is_make_fx_tracing requries this separate condition.

ghstack-source-id: f9ae2d5b631b719e1b3df10d71c18f12949635ec
Pull Request resolved: https://github.com/pytorch/pytorch/pull/164721